requirements.txt
fastapi~=0.61.2
uvicorn~=0.12.3
tortoise-orm[aiomysql]~=0.16.18
aiosqlite~=0.16.0
selenium~=3.141.0
requests~=2.25.0
pydantic~=1.7.2
invoke~=1.4.1
yarl~=1.6.3
PyYAML~=5.3.1
fire~=0.3.1
fabric~=2.5.0
aerich~=0.5.0
python-jose[cryptography]~=3.2.0
passlib[bcrypt]~=1.7.4